+++ core/vm/lfvm/interpreter.go
@@ -238,7 +238,7 @@ func runWithShadowInterpreter(c *context) {
 			fmt.Printf("Right: %v\n", *c.interpreter.Stack.Back(0))
 			panic("Stack top value divereged!")
 		}
-		if c.memory.Len() != uint(c.interpreter.Memory.Len()) {
+		if c.memory.Len() != uint64(c.interpreter.Memory.Len()) {
 			fmt.Printf("Left:  %v\n", c.memory.Len())
 			fmt.Printf("Right: %v\n", c.interpreter.Memory.Len())
 			panic("Memory size divereged!")
